Split out oracle.py (querying cards) and multiverse.py (building the db) into separate files.  Fixes #340
